Move zebra-state tests into their own files
