+++ zebra-state/src/service/non_finalized_state/chain.rs
@@ -403,203 +403,3 @@ impl Ord for Chain {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use proptest::{
-        num::usize::BinarySearch,
-        strategy::{NewTree, ValueTree},
-        test_runner::TestRunner,
-    };
-    use std::{env, sync::Arc};
-
-    use zebra_chain::{
-        block::Block,
-        parameters::{Network, NetworkUpgrade},
-        serialization::ZcashDeserializeInto,
-        LedgerState,
-    };
-    use zebra_test::prelude::*;
-
-    use crate::tests::{FakeChainHelper, Prepare};
-
-    use self::assert_eq;
-    use super::*;
-
-    const MAX_PARTIAL_CHAIN_BLOCKS: usize = 100;
-    const DEFAULT_PARTIAL_CHAIN_PROPTEST_CASES: u32 = 32;
-
-    #[test]
-    fn construct_empty() {
-        zebra_test::init();
-        let _chain = Chain::default();
-    }
-
-    #[test]
-    fn construct_single() -> Result<()> {
-        zebra_test::init();
-        let block: Arc<Block> =
-            zebra_test::vectors::BLOCK_MAINNET_434873_BYTES.zcash_deserialize_into()?;
-
-        let mut chain = Chain::default();
-        chain.push(block.prepare());
-
-        assert_eq!(1, chain.blocks.len());
-
-        Ok(())
-    }
-
-    #[test]
-    fn construct_many() -> Result<()> {
-        zebra_test::init();
-
-        let mut block: Arc<Block> =
-            zebra_test::vectors::BLOCK_MAINNET_434873_BYTES.zcash_deserialize_into()?;
-        let mut blocks = vec![];
-
-        while blocks.len() < 100 {
-            let next_block = block.make_fake_child();
-            blocks.push(block);
-            block = next_block;
-        }
-
-        let mut chain = Chain::default();
-
-        for block in blocks {
-            chain.push(block.prepare());
-        }
-
-        assert_eq!(100, chain.blocks.len());
-
-        Ok(())
-    }
-
-    #[test]
-    fn ord_matches_work() -> Result<()> {
-        zebra_test::init();
-        let less_block = zebra_test::vectors::BLOCK_MAINNET_434873_BYTES
-            .zcash_deserialize_into::<Arc<Block>>()?
-            .set_work(1);
-        let more_block = less_block.clone().set_work(10);
-
-        let mut lesser_chain = Chain::default();
-        lesser_chain.push(less_block.prepare());
-
-        let mut bigger_chain = Chain::default();
-        bigger_chain.push(more_block.prepare());
-
-        assert!(bigger_chain > lesser_chain);
-
-        Ok(())
-    }
-
-    #[derive(Debug)]
-    struct PreparedChainTree {
-        chain: Arc<Vec<PreparedBlock>>,
-        count: BinarySearch,
-    }
-
-    impl ValueTree for PreparedChainTree {
-        type Value = (Arc<Vec<PreparedBlock>>, <BinarySearch as ValueTree>::Value);
-
-        fn current(&self) -> Self::Value {
-            (self.chain.clone(), self.count.current())
-        }
-
-        fn simplify(&mut self) -> bool {
-            self.count.simplify()
-        }
-
-        fn complicate(&mut self) -> bool {
-            self.count.complicate()
-        }
-    }
-
-    #[derive(Debug, Default)]
-    struct PreparedChain {
-        // the proptests are threaded (not async), so we want to use a threaded mutex here
-        chain: std::sync::Mutex<Option<Arc<Vec<PreparedBlock>>>>,
-    }
-
-    impl Strategy for PreparedChain {
-        type Tree = PreparedChainTree;
-        type Value = <PreparedChainTree as ValueTree>::Value;
-
-        fn new_tree(&self, runner: &mut TestRunner) -> NewTree<Self> {
-            let mut chain = self.chain.lock().unwrap();
-            if chain.is_none() {
-                let tip_height = NetworkUpgrade::Canopy
-                    .activation_height(Network::Mainnet)
-                    .unwrap();
-                let ledger_state = LedgerState::new(tip_height, Network::Mainnet);
-                let blocks = Block::partial_chain_strategy(ledger_state, MAX_PARTIAL_CHAIN_BLOCKS)
-                    .prop_map(|vec| vec.into_iter().map(|blk| blk.prepare()).collect::<Vec<_>>())
-                    .new_tree(runner)?
-                    .current();
-                *chain = Some(Arc::new(blocks));
-            }
-
-            let chain = chain.clone().expect("should be generated");
-            let count = (1..chain.len()).new_tree(runner)?;
-            Ok(PreparedChainTree { chain, count })
-        }
-    }
-
-    #[test]
-    fn forked_equals_pushed() -> Result<()> {
-        zebra_test::init();
-
-        proptest!(ProptestConfig::with_cases(env::var("PROPTEST_CASES")
-                                          .ok()
-                                          .and_then(|v| v.parse().ok())
-                                          .unwrap_or(DEFAULT_PARTIAL_CHAIN_PROPTEST_CASES)),
-        |((chain, count) in PreparedChain::default())| {
-            let fork_tip_hash = chain[count - 1].hash;
-            let mut full_chain = Chain::default();
-            let mut partial_chain = Chain::default();
-
-            for block in chain.iter().take(count) {
-                partial_chain.push(block.clone());
-            }
-            for block in chain.iter() {
-                full_chain.push(block.clone());
-            }
-
-            let forked = full_chain.fork(fork_tip_hash).expect("hash is present");
-
-            prop_assert_eq!(forked.blocks.len(), partial_chain.blocks.len());
-        });
-
-        Ok(())
-    }
-
-    #[test]
-    fn finalized_equals_pushed() -> Result<()> {
-        zebra_test::init();
-
-        proptest!(ProptestConfig::with_cases(env::var("PROPTEST_CASES")
-                                          .ok()
-                                          .and_then(|v| v.parse().ok())
-                                          .unwrap_or(DEFAULT_PARTIAL_CHAIN_PROPTEST_CASES)),
-        |((chain, end_count) in PreparedChain::default())| {
-            let finalized_count = chain.len() - end_count;
-            let mut full_chain = Chain::default();
-            let mut partial_chain = Chain::default();
-
-            for block in chain.iter().skip(finalized_count) {
-                partial_chain.push(block.clone());
-            }
-            for block in chain.iter() {
-                full_chain.push(block.clone());
-            }
-
-            for _ in 0..finalized_count {
-                let _finalized = full_chain.pop_root();
-            }
-
-            prop_assert_eq!(full_chain.blocks.len(), partial_chain.blocks.len());
-        });
-
-        Ok(())
-    }
-}
