+++ zebra-state/src/service/non_finalized_state/tests.rs
@@ -1,0 +1,2 @@
+mod prop;
+mod vectors;
